+++ src/digital/mod.rs
@@ -1,6 +1,13 @@
 //! Digital I/O
+//!
+//! The traits in this module are now deprecated. Please use the new versions included
+//! in `digital::v2`.
 
 /// Single digital push-pull output pin
+///
+/// *This version of the trait is now deprecated. Please use the new `OutputPin` trait in
+/// `digital::v2::OutputPin`*.
+#[deprecated]
 pub trait OutputPin {
     /// Drives the pin low
     ///
@@ -18,6 +25,10 @@ pub trait OutputPin {
 /// Push-pull output pin that can read its output state
 ///
 /// *This trait is available if embedded-hal is built with the `"unproven"` feature.*
+///
+/// *This version of the trait is now deprecated. Please use the new `StatefulOutputPin` trait in
+/// `digital::v2::StatefulOutputPin`*.
+#[deprecated]
 #[cfg(feature = "unproven")]
 pub trait StatefulOutputPin {
     /// Is the pin in drive high mode?
@@ -35,10 +46,14 @@ pub trait StatefulOutputPin {
 ///
 /// *This trait is available if embedded-hal is built with the `"unproven"` feature.*
 ///
+/// *This version of the trait is now deprecated. Please use the new `ToggleableOutputPin`
+/// trait in `digital::v2::ToggleableOutputPin`*.
+///
 /// See [toggleable](toggleable) to use a software implementation if
 /// both [OutputPin](trait.OutputPin.html) and
 /// [StatefulOutputPin](trait.StatefulOutputPin.html) are
 /// implemented. Otherwise, implement this using hardware mechanisms.
+#[deprecated]
 #[cfg(feature = "unproven")]
 pub trait ToggleableOutputPin {
     /// Toggle pin output.
@@ -47,6 +62,9 @@ pub trait ToggleableOutputPin {
 
 /// If you can read **and** write the output state, a pin is
 /// toggleable by software.
+///
+/// *This version of the module is now deprecated. Please use the new `toggleable` module in
+/// `digital::v2::toggleable`*.
 ///
 /// ```
 /// use embedded_hal::digital::{OutputPin, StatefulOutputPin, ToggleableOutputPin};
@@ -84,15 +102,19 @@ pub trait ToggleableOutputPin {
 /// pin.toggle();
 /// assert!(pin.is_set_low());
 /// ```
+#[deprecated]
 #[cfg(feature = "unproven")]
 pub mod toggleable {
+    #[allow(deprecated)]
     use super::{OutputPin, StatefulOutputPin, ToggleableOutputPin};
 
     /// Software-driven `toggle()` implementation.
     ///
     /// *This trait is available if embedded-hal is built with the `"unproven"` feature.*
+    #[allow(deprecated)]
     pub trait Default: OutputPin + StatefulOutputPin {}
 
+    #[allow(deprecated)]
     impl<P> ToggleableOutputPin for P
     where
         P: Default,
@@ -111,6 +133,10 @@ pub mod toggleable {
 /// Single digital input pin
 ///
 /// *This trait is available if embedded-hal is built with the `"unproven"` feature.*
+///
+/// *This version of the trait is now deprecated. Please use the new `InputPin` trait in
+/// `digital::v2::InputPin`*.
+#[deprecated]
 #[cfg(feature = "unproven")]
 pub trait InputPin {
     /// Is the input pin high?
@@ -119,3 +145,6 @@ pub trait InputPin {
     /// Is the input pin low?
     fn is_low(&self) -> bool;
 }
+
+/// Improved version of the digital traits where the methods can also return an error.
+pub mod v2;